third_party/WebKit/LayoutTests/media/fullscreen-controls-visible-last.html

<!DOCTYPE html>
<title>Tests that the fullscreen button control is visible last.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<video controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "content/test");
    video.setAttribute("width", "70");
    video.addTextTrack("captions");
    video.onloadeddata = t.step_func_done(function() {
        assert_true(hasFullscreenButton(video));
        assert_false(isClosedCaptionsButtonVisible(video));
    });
});
</script>
